backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
alembic==1.12.0
python-dotenv==1.0.0
bcrypt==4.0.1
gspread==5.11.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
requests==2.31.0
Werkzeug==2.3.7
python-dateutil==2.8.2
pytz==2023.3